tlbuswrapper: CanAttachTLSlaves and CanAttachTLMasters
