+++ src/main/scala/subsystem/PeripheryBus.scala
@@ -19,76 +19,12 @@ case class PeripheryBusParams(
 
 case object PeripheryBusKey extends Field[PeripheryBusParams]
 
-class PeripheryBus(params: PeripheryBusParams)
-                  (implicit p: Parameters) extends TLBusWrapper(params, "periphery_bus")
+class PeripheryBus(params: PeripheryBusParams)(implicit p: Parameters)
+    extends TLBusWrapper(params, "periphery_bus")
+    with CanAttachTLSlaves
     with HasTLXbarPhy {
 
   val sbusXing = new CrossingHelper(this, params.sbusCrossingType)
-
-  def toSlave[D,U,E,B <: Data]
-      (name: Option[String] = None, buffer: BufferParams = BufferParams.none)
-      (gen: => NodeHandle[TLClientPortParameters,TLManagerPortParameters,TLEdgeIn,TLBundle,D,U,E,B] =
-        TLNameNode(name)): OutwardNodeHandle[D,U,E,B] = {
-    to("slave" named name) { gen :*= bufferTo(buffer) }
-  }
-
-  def toVariableWidthSlaveNode(name: Option[String] = None, buffer: BufferParams = BufferParams.none)(node: TLInwardNode) { toVariableWidthSlaveNodeOption(name, buffer)(Some(node)) }
-
-  def toVariableWidthSlaveNodeOption(name: Option[String] = None, buffer: BufferParams = BufferParams.none)(node: Option[TLInwardNode]) {
-    node foreach { n => to("slave" named name) { n :*= fragmentTo(buffer) } }
-  }
-
-  def toVariableWidthSlave[D,U,E,B <: Data]
-      (name: Option[String] = None, buffer: BufferParams = BufferParams.none)
-      (gen: => NodeHandle[TLClientPortParameters,TLManagerPortParameters,TLEdgeIn,TLBundle,D,U,E,B] =
-        TLNameNode(name)): OutwardNodeHandle[D,U,E,B] = {
-    to("slave" named name) { gen :*= fragmentTo(buffer) }
-  }
-
-  def toFixedWidthSlaveNode(name: Option[String] = None, buffer: BufferParams = BufferParams.none)(gen: TLInwardNode) {
-    to("slave" named name) { gen :*= fixedWidthTo(buffer) }
-  }
-
-  def toFixedWidthSlave[D,U,E,B <: Data]
-      (name: Option[String] = None, buffer: BufferParams = BufferParams.none)
-      (gen: => NodeHandle[TLClientPortParameters,TLManagerPortParameters,TLEdgeIn,TLBundle,D,U,E,B] =
-        TLNameNode(name)): OutwardNodeHandle[D,U,E,B] = {
-    to("slave" named name) { gen :*= fixedWidthTo(buffer) }
-  }
-
-  def toFixedWidthSingleBeatSlaveNode
-      (widthBytes: Int, name: Option[String] = None, buffer: BufferParams = BufferParams.none)
-      (gen: TLInwardNode) {
-    to("slave" named name) {
-      gen :*= TLFragmenter(widthBytes, params.blockBytes) :*= fixedWidthTo(buffer)
-    }
-  }
-
-  def toFixedWidthSingleBeatSlave[D,U,E,B <: Data]
-      (widthBytes: Int, name: Option[String] = None, buffer: BufferParams = BufferParams.none)
-      (gen: => NodeHandle[TLClientPortParameters,TLManagerPortParameters,TLEdgeIn,TLBundle,D,U,E,B] =
-        TLNameNode(name)): OutwardNodeHandle[D,U,E,B] = {
-    to("slave" named name) {
-      gen :*= TLFragmenter(widthBytes, params.blockBytes) :*= fixedWidthTo(buffer)
-    }
-  }
-
-  def toLargeBurstSlave[D,U,E,B <: Data]
-      (maxXferBytes: Int, name: Option[String] = None, buffer: BufferParams = BufferParams.none)
-      (gen: => NodeHandle[TLClientPortParameters,TLManagerPortParameters,TLEdgeIn,TLBundle,D,U,E,B] =
-        TLNameNode(name)): OutwardNodeHandle[D,U,E,B] = {
-    to("slave" named name) {
-      gen :*= fragmentTo(params.beatBytes, maxXferBytes, buffer)
-    }
-  }
-
-  def toFixedWidthPort[D,U,E,B <: Data]
-      (name: Option[String] = None, buffer: BufferParams = BufferParams.none)
-      (gen: => NodeHandle[TLClientPortParameters,TLManagerPortParameters,TLEdgeIn,TLBundle,D,U,E,B] =
-        TLNameNode(name)): OutwardNodeHandle[D,U,E,B] = {
-    to("port" named name) { gen := fixedWidthTo(buffer) }
-  }
-
 
   def fromSystemBus(gen: => TLOutwardNode) {
     from("sbus") {
@@ -103,15 +39,16 @@ class PeripheryBus(params: PeripheryBusParams)
       (name: Option[String] = None, buffer: BufferParams = BufferParams.none)
       (gen: => NodeHandle[D,U,E,B,TLClientPortParameters,TLManagerPortParameters,TLEdgeOut,TLBundle] =
         TLNameNode(name)): InwardNodeHandle[D,U,E,B] = {
-    from("master" named name) { bufferFrom(buffer) :=* gen }
+    from("master" named name) {
+      inwardNode :=* TLBuffer(buffer) :=* gen
+    }
   }
-
 
   def toTile
       (name: Option[String] = None, buffer: BufferParams = BufferParams.none)
       (gen: => TLNode): TLOutwardNode = {
     to("tile" named name) { FlipRendering { implicit p =>
-      gen :*= bufferTo(buffer)
+      gen :*= TLBuffer(buffer) :*= outwardNode
     }}
   }
 }
