+++ src/main/scala/subsystem/SystemBus.scala
@@ -17,11 +17,11 @@ case class SystemBusParams(
 
 case object SystemBusKey extends Field[SystemBusParams]
 
-class SystemBus(params: SystemBusParams)(implicit p: Parameters) extends TLBusWrapper(params, "system_bus")
+class SystemBus(params: SystemBusParams)(implicit p: Parameters)
+    extends TLBusWrapper(params, "system_bus")
+    with CanAttachTLSlaves
+    with CanAttachTLMasters
     with HasTLXbarPhy {
-
-  private val master_splitter = LazyModule(new TLSplitter)
-  inwardNode :=* master_splitter.node
 
   val cbus_params = new PeripheryBusParams(
     p(PeripheryBusKey).beatBytes,
@@ -34,8 +34,8 @@ class SystemBus(params: SystemBusParams)(implicit p: Parameters) extends TLBusWr
     TLFIFOFixer(TLFIFOFixer.all) :*= TLWidthWidget(params.beatBytes) :*= bufferTo(params.pbusBuffer)
   }
 
-  protected def fixFromThenSplit(policy: TLFIFOFixer.Policy, buffer: BufferParams): TLInwardNode =
-    master_splitter.node :=* TLBuffer(buffer) :=* TLFIFOFixer(policy)
+  private val master_splitter = LazyModule(new TLSplitter)
+  inwardNode :=* master_splitter.node
 
   def busView = master_splitter.node.edges.in.head
 
@@ -44,7 +44,8 @@ class SystemBus(params: SystemBusParams)(implicit p: Parameters) extends TLBusWr
       (gen
         :*= TLFIFOFixer(TLFIFOFixer.all)
         :*= TLWidthWidget(params.beatBytes)
-        :*= bufferTo(params.pbusBuffer))
+        :*= TLBuffer(params.pbusBuffer)
+        :*= outwardNode)
     }
   }
 
@@ -52,32 +53,11 @@ class SystemBus(params: SystemBusParams)(implicit p: Parameters) extends TLBusWr
     to("mbus") { gen := outwardNode }
   }
 
-  def toSlave[D,U,E,B <: Data]
-      (name: Option[String] = None, buffer: BufferParams = BufferParams.default)
-      (gen: => NodeHandle[TLClientPortParameters,TLManagerPortParameters,TLEdgeIn,TLBundle,D,U,E,B] =
-        TLNameNode(name)): OutwardNodeHandle[D,U,E,B] = {
-    to("slave" named name) { gen :*= bufferTo(buffer) }
-  }
- 
   def toSplitSlave[D,U,E,B <: Data]
       (name: Option[String] = None)
       (gen: => NodeHandle[TLClientPortParameters,TLManagerPortParameters,TLEdgeIn,TLBundle,D,U,E,B] =
         TLNameNode(name)): OutwardNodeHandle[D,U,E,B] = {
     to("slave" named name) { gen :=* master_splitter.node }
-  }
-
-  def toFixedWidthSlave[D,U,E,B <: Data]
-      (name: Option[String] = None, buffer: BufferParams = BufferParams.default)
-      (gen: =>  NodeHandle[TLClientPortParameters,TLManagerPortParameters,TLEdgeIn,TLBundle,D,U,E,B] =
-        TLNameNode(name)): OutwardNodeHandle[D,U,E,B] = {
-    to("slave" named name) { gen :*= fixedWidthTo(buffer) }
-  }
-
-  def toVariableWidthSlave[D,U,E,B <: Data]
-      (name: Option[String] = None, buffer: BufferParams = BufferParams.default)
-      (gen: => NodeHandle[TLClientPortParameters,TLManagerPortParameters,TLEdgeIn,TLBundle,D,U,E,B] =
-        TLNameNode(name)): OutwardNodeHandle[D,U,E,B] = {
-    to("slave" named name) { gen :*= fragmentTo(buffer) }
   }
 
   def fromFrontBus(gen: => TLNode): TLInwardNode = {
@@ -88,36 +68,7 @@ class SystemBus(params: SystemBusParams)(implicit p: Parameters) extends TLBusWr
       (name: Option[String], buffer: BufferParams = BufferParams.none, cork: Option[Boolean] = None)
       (gen: => TLNode): TLInwardNode = {
     from("tile" named name) {
-      fixFromThenSplit(TLFIFOFixer.allUncacheable, buffer) :=* gen
+      master_splitter.node :=* TLBuffer(buffer) :=* TLFIFOFixer(TLFIFOFixer.allUncacheable) :=* gen
     }
   }
-
-  def toFixedWidthPort[D,U,E,B <: Data]
-      (name: Option[String] = None, buffer: BufferParams = BufferParams.default)
-      (gen: => NodeHandle[TLClientPortParameters,TLManagerPortParameters,TLEdgeIn,TLBundle,D,U,E,B] =
-        TLNameNode(name)): OutwardNodeHandle[D,U,E,B] = {
-    to("port" named name) { gen := fixedWidthTo(buffer) }
-  }
-
-  def fromPort[D,U,E,B <: Data]
-      (name: Option[String] = None, buffer: BufferParams = BufferParams.none)
-      (gen: => NodeHandle[D,U,E,B,TLClientPortParameters,TLManagerPortParameters,TLEdgeOut,TLBundle] =
-        TLNameNode(name)): InwardNodeHandle[D,U,E,B] = {
-    from("port" named name) { fixFromThenSplit(TLFIFOFixer.all, buffer) :=* gen }
-  }
-
-  def fromCoherentMaster[D,U,E,B <: Data]
-      (name: Option[String] = None, buffer: BufferParams = BufferParams.none)
-      (gen: => NodeHandle[D,U,E,B,TLClientPortParameters,TLManagerPortParameters,TLEdgeOut,TLBundle] =
-        TLNameNode(name)): InwardNodeHandle[D,U,E,B] = {
-    from("coherent_master" named name) { fixFrom(TLFIFOFixer.all, buffer) :=* gen }
-  }
-
-  def fromMaster[D,U,E,B <: Data]
-      (name: Option[String] = None, buffer: BufferParams = BufferParams.none)
-      (gen: => NodeHandle[D,U,E,B,TLClientPortParameters,TLManagerPortParameters,TLEdgeOut,TLBundle] =
-        TLNameNode(name)): InwardNodeHandle[D,U,E,B] = {
-    from("master" named name) { fixFromThenSplit(TLFIFOFixer.all, buffer) :=* gen }
-  }
-
 }
