+++ src/main/scala/tilelink/BusWrapper.scala
@@ -5,6 +5,7 @@ package freechips.rocketchip.tilelink
 import Chisel._
 import freechips.rocketchip.config.Parameters
 import freechips.rocketchip.diplomacy._
+import freechips.rocketchip.util._
 
 /** Specifies widths of various attachement points in the SoC */
 trait HasTLBusParams {
@@ -18,7 +19,9 @@ trait HasTLBusParams {
 }
 
 abstract class TLBusWrapper(params: HasTLBusParams, val busName: String)(implicit p: Parameters)
-    extends SimpleLazyModule with LazyScope with HasTLBusParams {
+    extends SimpleLazyModule
+    with LazyScope
+    with HasTLBusParams {
 
   val beatBytes = params.beatBytes
   val blockBytes = params.blockBytes
@@ -51,6 +54,103 @@ abstract class TLBusWrapper(params: HasTLBusParams, val busName: String)(implici
 
   def from[T](name: String)(body: => T): T = {
     this { LazyScope(s"coupler_from_${name}") { body } }
+  }
+}
+
+trait CanAttachTLSlaves { this: TLBusWrapper =>
+  def toSlave[D,U,E,B <: Data]
+      (name: Option[String] = None, buffer: BufferParams = BufferParams.none)
+      (gen: => NodeHandle[TLClientPortParameters,TLManagerPortParameters,TLEdgeIn,TLBundle,D,U,E,B] =
+        TLNameNode(name)): OutwardNodeHandle[D,U,E,B] = {
+    to("slave" named name) { gen :*= bufferTo(buffer) }
+  }
+
+  def toVariableWidthSlaveNode(name: Option[String] = None, buffer: BufferParams = BufferParams.none)(node: TLInwardNode) {
+    toVariableWidthSlaveNodeOption(name, buffer)(Some(node))
+  }
+
+  def toVariableWidthSlaveNodeOption(name: Option[String] = None, buffer: BufferParams = BufferParams.none)(node: Option[TLInwardNode]) {
+    node foreach { n => to("slave" named name) { n :*= fragmentTo(buffer) } }
+  }
+
+  def toVariableWidthSlave[D,U,E,B <: Data]
+      (name: Option[String] = None, buffer: BufferParams = BufferParams.none)
+      (gen: => NodeHandle[TLClientPortParameters,TLManagerPortParameters,TLEdgeIn,TLBundle,D,U,E,B] =
+        TLNameNode(name)): OutwardNodeHandle[D,U,E,B] = {
+    to("slave" named name) { gen :*= fragmentTo(buffer) }
+  }
+
+  def toFixedWidthSlaveNode(name: Option[String] = None, buffer: BufferParams = BufferParams.none)(gen: TLInwardNode) {
+    to("slave" named name) { gen :*= fixedWidthTo(buffer) }
+  }
+
+  def toFixedWidthSlave[D,U,E,B <: Data]
+      (name: Option[String] = None, buffer: BufferParams = BufferParams.none)
+      (gen: => NodeHandle[TLClientPortParameters,TLManagerPortParameters,TLEdgeIn,TLBundle,D,U,E,B] =
+        TLNameNode(name)): OutwardNodeHandle[D,U,E,B] = {
+    to("slave" named name) { gen :*= fixedWidthTo(buffer) }
+  }
+
+  def toFixedWidthSingleBeatSlaveNode
+      (widthBytes: Int, name: Option[String] = None, buffer: BufferParams = BufferParams.none)
+      (gen: TLInwardNode) {
+    to("slave" named name) {
+      gen :*= TLFragmenter(widthBytes, blockBytes) :*= fixedWidthTo(buffer)
+    }
+  }
+
+  def toFixedWidthSingleBeatSlave[D,U,E,B <: Data]
+      (widthBytes: Int, name: Option[String] = None, buffer: BufferParams = BufferParams.none)
+      (gen: => NodeHandle[TLClientPortParameters,TLManagerPortParameters,TLEdgeIn,TLBundle,D,U,E,B] =
+        TLNameNode(name)): OutwardNodeHandle[D,U,E,B] = {
+    to("slave" named name) {
+      gen :*= TLFragmenter(widthBytes, blockBytes) :*= fixedWidthTo(buffer)
+    }
+  }
+
+  def toLargeBurstSlave[D,U,E,B <: Data]
+      (maxXferBytes: Int, name: Option[String] = None, buffer: BufferParams = BufferParams.none)
+      (gen: => NodeHandle[TLClientPortParameters,TLManagerPortParameters,TLEdgeIn,TLBundle,D,U,E,B] =
+        TLNameNode(name)): OutwardNodeHandle[D,U,E,B] = {
+    to("slave" named name) {
+      gen :*= fragmentTo(beatBytes, maxXferBytes, buffer)
+    }
+  }
+
+  def toFixedWidthPort[D,U,E,B <: Data]
+      (name: Option[String] = None, buffer: BufferParams = BufferParams.none)
+      (gen: => NodeHandle[TLClientPortParameters,TLManagerPortParameters,TLEdgeIn,TLBundle,D,U,E,B] =
+        TLNameNode(name)): OutwardNodeHandle[D,U,E,B] = {
+    to("port" named name) { gen := fixedWidthTo(buffer) }
+  }
+}
+
+trait CanAttachTLMasters { this: TLBusWrapper =>
+  def fromMasterNode
+      (name: Option[String] = None, buffer: BufferParams = BufferParams.none)
+      (gen: TLOutwardNode) {
+    from("master" named name) { fixFrom(TLFIFOFixer.all, buffer) :=* gen }
+  }
+
+  def fromMaster[D,U,E,B <: Data]
+      (name: Option[String] = None, buffer: BufferParams = BufferParams.none)
+      (gen: => NodeHandle[D,U,E,B,TLClientPortParameters,TLManagerPortParameters,TLEdgeOut,TLBundle] =
+        TLNameNode(name)): InwardNodeHandle[D,U,E,B] = {
+    from("master" named name) { fixFrom(TLFIFOFixer.all, buffer) :=* gen }
+  }
+
+  def fromPort[D,U,E,B <: Data]
+      (name: Option[String] = None, buffer: BufferParams = BufferParams.none)
+      (gen: => NodeHandle[D,U,E,B,TLClientPortParameters,TLManagerPortParameters,TLEdgeOut,TLBundle] =
+        TLNameNode(name)): InwardNodeHandle[D,U,E,B] = {
+    from("port" named name) { fixFrom(TLFIFOFixer.all, buffer) :=* gen }
+  }
+
+  def fromCoherentMaster[D,U,E,B <: Data]
+      (name: Option[String] = None, buffer: BufferParams = BufferParams.none)
+      (gen: => NodeHandle[D,U,E,B,TLClientPortParameters,TLManagerPortParameters,TLEdgeOut,TLBundle] =
+        TLNameNode(name)): InwardNodeHandle[D,U,E,B] = {
+    from("coherent_master" named name) { fixFrom(TLFIFOFixer.all, buffer) :=* gen }
   }
 }
